pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "engram"
version = "0.1.16"
description = "Local-first memory runtime for LLM applications with layered memory and a sandbox reference app"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "Jeffrey Dean"}]
readme = "README.md"

dependencies = [
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0",
]


[project.scripts]
engram = "engram.cli:main"

[project.optional-dependencies]
# Neural memory / RTRL (optional; keeps base install lightweight)
neural = [
    "torch>=2.0.0",
]
# Engine backends
engines = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "tenacity>=8.0.0",
    "tiktoken>=0.5.0",
    "json-repair>=0.7.0",
]
# Episodic memory (ChromaDB + embeddings)
episodic = [
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.0",
    "diskcache>=5.6.2",
]
# Semantic memory (Kuzu graph DB)
semantic = [
    "kuzu>=0.4.0",
]
# UI (Streamlit sandbox)
ui = [
    "streamlit>=1.30.0",
    "platformdirs>=4.0.0",
]

# Fine-tuning data export (no extra deps — uses stdlib + existing numpy/pydantic)
finetune = []

# Everything
all = [
    "engram[engines,episodic,semantic,neural,ui,finetune]",
]
# Development
dev = [
    "engram[all]",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
include = ["engram*", "engram_ui*", "apps*"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests", "tests/harness"]
pythonpath = ["."]
asyncio_mode = "auto"
# Suppress the SkipTest noise from dependency guards
filterwarnings = ["ignore::pytest.PytestUnraisableExceptionWarning"]
